Parsing expressions inside string literals (#56)
PHP supports expressions inside strings ("hello, $name" or
"{$a->b["aaa"]}"). This is an attempt to parse such expressions.  It
works but it's not completely compatible with Zend PHP (which further
restricts kinds of expressions that can be placed inside a string
literal). It also doesn't support expressions inside heredocs (Zend
does).