Warn about 'case' fall-through (#68)
Emit issues when a 'case' block passes control to the next one. If this
happens in the last 'case' it's potentially dangerous so a different
kind of issue is emitted.

This is somewhat over-zealous. In particular, it emits issues for fairly
common constructs like:

switch(...) {
case 0:
case 1:
...
}

A test case was added to solve that in the future.